Extract Tree utils from Reflection
* Allows TreeMap, TreeTraverser and TreeAccumulator to be extended in non local classes
* Removes logic that does not belong in Reflection
