+++ library/src/scala/tasty/reflect/TreeTraverser.scala
@@ -1,0 +1,25 @@
+package scala.tasty
+package reflect
+
+/** TASTy Reflect tree traverser.
+ *
+ *  Usage:
+ *  ```
+ *  class MyTraverser[R <: scala.tasty.Reflection & Singleton](val reflect: R)
+ *      extends scala.tasty.reflect.TreeTraverser {
+ *    import reflect.{given, _}
+ *    override def traverseTree(tree: Tree)(using ctx: Context): Unit = ...
+ *  }
+ *  ```
+ */
+trait TreeTraverser extends TreeAccumulator[Unit] {
+
+  import reflect._
+
+  def traverseTree(tree: Tree)(given ctx: Context): Unit = traverseTreeChildren(tree)
+
+  def foldTree(x: Unit, tree: Tree)(given ctx: Context): Unit = traverseTree(tree)
+
+  protected def traverseTreeChildren(tree: Tree)(given ctx: Context): Unit = foldOverTree((), tree)
+
+}
