+++ tests/run-custom-args/Yretain-trees/tasty-extractors-owners/quoted_1.scala
@@ -11,28 +11,31 @@ object Macros {
 
     val buff = new StringBuilder
 
-    val output = new TreeTraverser {
-      override def traverseTree(tree: Tree)(implicit ctx: Context): Unit = {
-        tree match {
-          case tree @ DefDef(name, _, _, _, _) =>
-            buff.append(name)
-            buff.append("\n")
-            buff.append(tree.symbol.owner.tree.showExtractors)
-            buff.append("\n\n")
-          case tree @ ValDef(name, _, _) =>
-            buff.append(name)
-            buff.append("\n")
-            buff.append(tree.symbol.owner.tree.showExtractors)
-            buff.append("\n\n")
-          case _ =>
-        }
-        traverseTreeChildren(tree)
-      }
-    }
+    val output = new MyTraverser(qctx.tasty)(buff)
 
     val tree = x.unseal
     output.traverseTree(tree)
     '{print(${buff.result()})}
+  }
+
+  class MyTraverser[R <: scala.tasty.Reflection & Singleton](val reflect: R)(buff: StringBuilder) extends scala.tasty.reflect.TreeTraverser {
+    import reflect.{given, _}
+    override def traverseTree(tree: Tree)(implicit ctx: Context): Unit = {
+      tree match {
+        case tree @ DefDef(name, _, _, _, _) =>
+          buff.append(name)
+          buff.append("\n")
+          buff.append(tree.symbol.owner.tree.showExtractors)
+          buff.append("\n\n")
+        case tree @ ValDef(name, _, _) =>
+          buff.append(name)
+          buff.append("\n")
+          buff.append(tree.symbol.owner.tree.showExtractors)
+          buff.append("\n\n")
+        case _ =>
+      }
+      traverseTreeChildren(tree)
+    }
   }
 
 }
