+++ tests/run-custom-args/tasty-inspector/tasty-comment-inspector/Test.scala
@@ -1,4 +1,5 @@
 import scala.tasty.Reflection
+import scala.tasty.reflect.TreeTraverser
 import scala.tasty.inspector._
 
 object Test {
@@ -9,10 +10,10 @@ object Test {
 
 class CommentInspector extends TastyInspector {
 
-  def processCompilationUnit(reflect: Reflection)(root: reflect.Tree): Unit = {
-    import reflect.{_, given _}
+  def processCompilationUnit(reflect0: Reflection)(root: reflect0.Tree): Unit = {
+    import reflect0.{_, given _}
     object Traverser extends TreeTraverser {
-
+      val reflect: reflect0.type = reflect0
       override def traverseTree(tree: Tree)(implicit ctx: Context): Unit = tree match {
         case tree: Definition =>
           tree.symbol.comment match {
@@ -25,7 +26,7 @@ class CommentInspector extends TastyInspector {
       }
 
     }
-    Traverser.traverseTree(root)(reflect.rootContext)
+    Traverser.traverseTree(root)(reflect0.rootContext)
   }
 
 }
