+++ tests/run-macros/tasty-extractors-3/quoted_1.scala
@@ -11,7 +11,8 @@ object Macros {
     import qctx.tasty.{_, given _}
 
     val buff = new StringBuilder
-    val traverser = new TreeTraverser {
+    val traverser = new scala.tasty.reflect.TreeTraverser {
+      val reflect: qctx.tasty.type = qctx.tasty
       override def traverseTree(tree: Tree)(implicit ctx: Context): Unit = tree match {
         case tree: TypeBoundsTree =>
           buff.append(tree.tpe.showExtractors)
